% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{scale_fill_seaborn}
\alias{scale_fill_seaborn}
\alias{scale_colour_seaborn}
\title{seaborn scales for ggplot2}
\usage{
scale_fill_seaborn(..., palette = "muted", aesthetics = "fill", direction = 1)

scale_colour_seaborn(
  ...,
  palette = "muted",
  aesthetics = "colour",
  direction = 1
)
}
\arguments{
\item{...}{passed to \link[ggplot2]{discrete_scale}}

\item{palette}{optional, name of palette, default is "muted". Use
\link{show_seaborn} to see all options}

\item{aesthetics}{optional, define which aesthetics (fill, colour)
to set the scale for}

\item{direction}{optional, default 1. set direction=-1 to use the
colours in the opposite order}
}
\value{
palette for ggplot2 object
}
\description{
seaborn scales for ggplot2
}
\details{
designed to bring the seaborn scale
}
\examples{
library(ggplot2)
ggplot(data.frame(x=1:5,y=1:5,fill=factor(1:5)), aes(x=x,y=y,fill=fill)) +
geom_tile()

ggplot(data.frame(x=1:5,y=1:5,fill=factor(1:5)), aes(x=x,y=y,fill=fill)) +
geom_tile() +
scale_fill_seaborn()

ggplot(data.frame(x=1:5,y=1:5,fill=factor(1:5)), aes(x=x,y=y,fill=fill)) +
geom_tile() +
scale_fill_seaborn(palette="dark6",direction=-1)
}
\author{
Chris Wallace
}
