% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{show_seaborn}
\alias{show_seaborn}
\title{show all seaborn palettes}
\usage{
show_seaborn()
}
\value{
ggplot2 object
}
\description{
display seaborn palettes
}
\details{
generate a display of all seaborn palette options on current graphics device
}
\examples{
show_seaborn()
}
\author{
Chris Wallace
}
